# refactor: remove ramda from textkit package