--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -28,6 +28,7 @@ module.exports = {
     'prefer-destructuring': 1,
     'no-use-before-define': 1,
     'no-underscore-dangle': 0,
+    'prefer-object-spread': 0,
     'import/no-named-as-default': 0,
     'react/state-in-constructor': 0,
     'react/jsx-filename-extension': 0,