--- a/packages/fns/dropLast.js
+++ b/packages/fns/dropLast.js
@@ -1,0 +1,3 @@
+const dropLast = array => array.slice(0, array.length - 1);
+
+export default dropLast;