--- a/packages/fns/isNil.js
+++ b/packages/fns/isNil.js
@@ -1,0 +1,3 @@
+const isNil = value => value === null || value === undefined;
+
+export default isNil;