--- a/packages/fns/last.js
+++ b/packages/fns/last.js
@@ -1,0 +1,5 @@
+const last = value => {
+  return value === '' ? '' : value[value.length - 1];
+};
+
+export default last;