--- a/packages/fns/omit.js
+++ b/packages/fns/omit.js
@@ -1,0 +1,9 @@
+const omit = (key, object) => {
+  const copy = Object.assign({}, object);
+
+  delete copy[key];
+
+  return copy;
+};
+
+export default omit;