--- a/packages/layout/src/svg/layoutText.js
+++ b/packages/layout/src/svg/layoutText.js
@@ -2,11 +2,11 @@ import * as R from 'ramda';
 import * as P from '@react-pdf/primitives';
 import layoutEngine from '@react-pdf/textkit/lib/layout';
 import linebreaker from '@react-pdf/textkit/lib/engines/linebreaker';
-import AttributedString from '@react-pdf/textkit/lib/attributedString';
 import justification from '@react-pdf/textkit/lib/engines/justification';
 import scriptItemizer from '@react-pdf/textkit/lib/engines/scriptItemizer';
 import wordHyphenation from '@react-pdf/textkit/lib/engines/wordHyphenation';
 import decorationEngine from '@react-pdf/textkit/lib/engines/textDecoration';
+import fromFragments from '@react-pdf/textkit/lib/attributedString/fromFragments';
 
 import transformText from '../text/transformText';
 import fontSubstitution from '../text/fontSubstitution';
@@ -83,7 +83,7 @@ const getFragments = (fontStore, instance) => {
 };
 
 const getAttributedString = (fontStore, instance) =>
-  AttributedString.fromFragments(getFragments(fontStore, instance));
+  fromFragments(getFragments(fontStore, instance));
 
 const AlmostInfinity = 999999999999;