--- a/packages/layout/src/text/linesWidth.js
+++ b/packages/layout/src/text/linesWidth.js
@@ -1,4 +1,4 @@
-import AttributedString from '@react-pdf/textkit/lib/attributedString';
+import advanceWidth from '@react-pdf/textkit/lib/attributedString/advanceWidth';
 
 /**
  * Get lines width (if any)
@@ -8,9 +8,7 @@ import AttributedString from '@react-pdf/textkit/lib/attributedString';
  */
 const linesWidth = node => {
   if (!node.lines) return -1;
-  return Math.max(
-    ...node.lines.map(line => AttributedString.advanceWidth(line)),
-  );
+  return Math.max(...node.lines.map(line => advanceWidth(line)));
 };
 
 export default linesWidth;