--- a/packages/textkit/src/attributedString/append.js
+++ b/packages/textkit/src/attributedString/append.js
@@ -1,5 +1,4 @@
-import * as R from 'ramda';
-
+import last from '../../../fns/last';
 import emptyRun from '../run/empty';
 import appendToRun from '../run/append';
 import stringFromCodePoints from '../utils/stringFromCodePoints';
@@ -11,20 +10,16 @@ import stringFromCodePoints from '../utils/stringFromCodePoints';
  * @param {Object} attributed string
  * @return {Object} attributed string with new glyph
  */
-const append = (glyph, string) => {
-  const codePoints = R.propOr([], 'codePoints')(glyph);
+const append = (glyph, attributedString) => {
+  const codePoints = glyph?.codePoints || [];
+  const codePointsString = stringFromCodePoints(codePoints);
+  const string = attributedString.string + codePointsString;
 
-  return R.evolve({
-    string: R.concat(R.__, stringFromCodePoints(codePoints)),
-    runs: R.converge(R.concat, [
-      R.init,
-      R.compose(
-        R.unapply(R.identity),
-        v => appendToRun(glyph, v),
-        R.either(R.last, emptyRun),
-      ),
-    ]),
-  })(string);
+  const firstRuns = attributedString.runs.slice(0, -1);
+  const lastRun = last(attributedString.runs) || emptyRun();
+  const runs = firstRuns.concat(appendToRun(glyph, lastRun));
+
+  return Object.assign({}, attributedString, { string, runs });
 };
 
-export default R.curryN(2, append);
+export default append;