--- a/packages/textkit/src/attributedString/dropLast.js
+++ b/packages/textkit/src/attributedString/dropLast.js
@@ -1,6 +1,6 @@
-import * as R from 'ramda';
-
 import runDropLast from '../run/dropLast';
+import adjust from '../../../fns/adjust';
+import arrayDropLast from '../../../fns/dropLast';
 
 /**
  * Drop last glyph
@@ -8,10 +8,11 @@ import runDropLast from '../run/dropLast';
  * @param {Object} attributed string
  * @return {Object} attributed string with new glyph
  */
-const dropLast = string =>
-  R.evolve({
-    string: R.dropLast(1),
-    runs: R.adjust(-1, runDropLast),
-  })(string);
+const dropLast = attributeString => {
+  const string = arrayDropLast(attributeString.string);
+  const runs = adjust(-1, runDropLast, attributeString.runs);
+
+  return Object.assign({}, attributeString, { string, runs });
+};
 
 export default dropLast;