--- a/packages/textkit/src/attributedString/empty.js
+++ b/packages/textkit/src/attributedString/empty.js
@@ -1,10 +1,8 @@
-import * as R from 'ramda';
-
 /**
  * Returns empty attributed string
  *
  * @return {Object} empty attributed string
  */
-const empty = R.always({ string: '', runs: [] });
+const empty = () => ({ string: '', runs: [] });
 
 export default empty;