--- a/packages/textkit/src/attributedString/fromFragments.js
+++ b/packages/textkit/src/attributedString/fromFragments.js
@@ -1,5 +1,3 @@
-import * as R from 'ramda';
-
 /**
  * Create attributed string from text fragments
  *
@@ -8,22 +6,22 @@ import * as R from 'ramda';
  */
 const fromFragments = fragments => {
   let offset = 0;
+  let string = '';
+  const runs = [];
 
-  const getRuns = R.map(fragment => {
-    const run = {
+  fragments.forEach(fragment => {
+    string += fragment.string;
+
+    runs.push({
       start: offset,
       end: offset + fragment.string.length,
       attributes: fragment.attributes || {},
-    };
-    offset += fragment.string.length;
+    });
 
-    return run;
+    offset += fragment.string.length;
   });
 
-  return R.applySpec({
-    runs: getRuns,
-    string: R.o(R.join(''), R.pluck('string')),
-  })(fragments);
+  return { string, runs };
 };
 
 export default fromFragments;