--- a/packages/textkit/src/attributedString/insertGlyph.js
+++ b/packages/textkit/src/attributedString/insertGlyph.js
@@ -1,16 +1,8 @@
-import * as R from 'ramda';
-
 import append from './append';
-import copy from '../run/copy';
+import add from '../run/add';
 import insert from '../run/insert';
 import runIndexAt from './runIndexAt';
 import stringFromCodePoints from '../utils/stringFromCodePoints';
-
-const mapCond = conds => R.addIndex(R.map)(R.cond(conds));
-
-const idxEquals = idx => R.compose(R.equals(idx), R.nthArg(1));
-
-const idxGt = idx => R.compose(R.gt(R.__, idx), R.nthArg(1));
 
 /**
  * Insert glyph into attributed string
@@ -20,34 +12,28 @@ const idxGt = idx => R.compose(R.gt(R.__, idx), R.nthArg(1));
  * @param {Object} attributed string
  * @return {Object} attributed string with new glyph
  */
-const insertGlyph = (index, glyph, string) => {
-  const runIndex = runIndexAt(index, string);
+const insertGlyph = (index, glyph, attributedString) => {
+  const runIndex = runIndexAt(index, attributedString);
 
   // Add glyph to the end if run index invalid
-  if (runIndex === -1) {
-    return append(glyph, string);
-  }
+  if (runIndex === -1) return append(glyph, attributedString);
 
-  const codePoints = R.propOr([], 'codePoints')(glyph);
-  const incRange = R.add(R.length(codePoints));
+  const codePoints = glyph?.codePoints || [];
 
-  return R.evolve({
-    string: R.compose(
-      R.join(''),
-      R.insert(index, stringFromCodePoints(codePoints)),
-    ),
-    runs: mapCond([
-      [idxEquals(runIndex), run => insert(index - run.start, glyph, run)],
-      [
-        idxGt(runIndex),
-        R.evolve({
-          start: incRange,
-          end: incRange,
-        }),
-      ],
-      [R.T, copy],
-    ]),
-  })(string);
+  const string =
+    attributedString.string.slice(0, index) +
+    stringFromCodePoints(codePoints) +
+    attributedString.string.slice(index);
+
+  const runs = attributedString.runs.map((run, i) => {
+    if (i === runIndex) return insert(index - run.start, glyph, run);
+
+    if (i > runIndex) return add(codePoints.length, run);
+
+    return run;
+  });
+
+  return Object.assign({}, attributedString, { string, runs });
 };
 
-export default R.curryN(3, insertGlyph);
+export default insertGlyph;