--- a/packages/textkit/src/attributedString/length.js
+++ b/packages/textkit/src/attributedString/length.js
@@ -1,5 +1,3 @@
-import * as R from 'ramda';
-
 import start from './start';
 import end from './end';
 
@@ -9,6 +7,8 @@ import end from './end';
  * @param  {Object}  glyph string
  * @return {number} end
  */
-const length = R.converge(R.subtract, [end, start]);
+const length = attributedString => {
+  return end(attributedString) - start(attributedString);
+};
 
 export default length;