--- a/packages/textkit/src/block/height.js
+++ b/packages/textkit/src/block/height.js
@@ -1,11 +1,13 @@
-import * as R from 'ramda';
-
 /**
- * Get paragrpah block height
+ * Get paragraph block height
  *
  * @param  {Object}  paragraph block
  * @return {number} paragraph block height
  */
-const height = R.compose(R.sum, R.map(R.prop('height')), R.pluck('box'));
+const height = paragraph => {
+  return paragraph.reduce((acc, block) => acc + block.box.height, 0);
+};
+
+// R.compose(R.sum, R.map(R.prop('height')), R.pluck('box'));
 
 export default height;