--- a/packages/textkit/src/block/height.js
+++ b/packages/textkit/src/block/height.js
@@ -8,6 +8,4 @@ const height = paragraph => {
   return paragraph.reduce((acc, block) => acc + block.box.height, 0);
 };
 
-// R.compose(R.sum, R.map(R.prop('height')), R.pluck('box'));
-
 export default height;