--- a/packages/textkit/src/block/slice.js
+++ b/packages/textkit/src/block/slice.js
@@ -1,5 +1,3 @@
-import * as R from 'ramda';
-
 /**
  * Slice block lines
  *
@@ -7,6 +5,6 @@ import * as R from 'ramda';
  * @param  {Object}  paragraph block
  * @return {Object} sliced paragraph block
  */
-const sliceBlock = (lines, block) => R.slice(0, lines, block);
+const sliceBlock = (lines, block) => block.slice(0, lines);
 
 export default sliceBlock;