… .changeset/tough-cows-give.md
… packages/fns/adjust.test.js
… packages/fns/dropLast.test.js
… packages/fns/get.test.js
… packages/fns/isNil.test.js
… packages/fns/last.test.js
… packages/fns/omit.test.js
… packages/textkit/package.json
… packages/textkit/src/attributedString/reduce.js
… packages/textkit/src/block/index.js
… packages/textkit/src/engines/justification/getFactors.js
… packages/textkit/src/engines/justification/index.js
… packages/textkit/src/engines/linebreaker/index.js
… packages/textkit/src/engines/scriptItemizer/index.js
… packages/textkit/src/engines/textDecoration/index.js
… packages/textkit/src/engines/wordHyphenation/index.js
… packages/textkit/src/glyph/index.js
… packages/textkit/src/glyph/isWhiteSpace.js
… packages/textkit/src/glyph/slice.js
… packages/textkit/src/indices/append.js
… packages/textkit/src/indices/index.js
… packages/textkit/src/indices/normalize.js
… packages/textkit/src/indices/prepend.js
… packages/textkit/src/layout/applyDefaultStyles.js
… packages/textkit/src/layout/finalizeFragments.js
… packages/textkit/src/layout/generateGlyphs.js
… packages/textkit/src/layout/index.js
… packages/textkit/src/layout/layoutParagraph.js
… packages/textkit/src/layout/preprocessRuns.js
… packages/textkit/src/layout/resolveAttachments.js