## Summary

Gradually start removing ramda. Although I like it, it's probably confusing for most people, it's slow and adds unneccesary bundle size. Some helper fns will still be needed. For the moment I moved them to utils dir, in the future it worths building our own set of light and reusable utils

## Benchmarks

**Before**

```
generate doc x 36 ops/sec ±4.03% (68 runs sampled)
✨  Done in 7.71s.
```

**After**

```
generate doc x 67 ops/sec ±4.02% (73 runs sampled)
✨  Done in 8.07s.
```

This is going to be huge 😄